% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaria_trial_pk.R
\docType{package}
\name{malariatrialpk}
\alias{malariatrialpk}
\title{Estimates pharmacokinetic parameters from trial data via MCMC}
\description{
Note, this package was forked from the MCMC package drjacoby, so
  there may be some names throughout that come from there.
}
