% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run drjacoby MCMC}
\usage{
run_mcmc(
  data,
  burnin = 1000,
  samples = 10000,
  chains = 5,
  beta_manual = 1,
  target_acceptance = 0.44,
  coupling_on = TRUE,
  pb_markdown = FALSE,
  silent = FALSE
)
}
\arguments{
\item{data}{a named list of numeric data values. When using C++ likelihood
and/or prior these values are treated internally as doubles, so while
integer and Boolean values can be used, keep in mind that these will be
recast as doubles in the likelihood (i.e. \code{TRUE = 1.0}).}

\item{burnin}{the number of burn-in iterations. Automatic tuning of proposal
standard deviations is only active during the burn-in period.}

\item{samples}{the number of sampling iterations.}

\item{chains}{the number of independent replicates of the MCMC to run.}

\item{beta_manual}{vector of manually defined \eqn{\beta} values used in the
parallel tempering approach. If defined, this overrides the spacing defined
by \code{rungs}. Note that even manually defined \eqn{\beta} values are
raised to the power \eqn{\alpha} internally, hence you should set
\code{alpha = 1} if you want to fix \eqn{\beta} values exactly.}

\item{target_acceptance}{Target acceptance rate. Should be between 0 and 1.
Default of 0.44, set as optimum for unvariate proposal distributions.}

\item{coupling_on}{whether to implement Metropolis-coupling over temperature
rungs. The option of deactivating coupling has been retained for general
interest and debugging purposes only. If this parameter is \code{FALSE}
then parallel tempering will have no impact on MCMC mixing.}

\item{pb_markdown}{whether to run progress bars in markdown mode, meaning
they are only updated when they reach 100\% to avoid large amounts of output
being printed to markdown files.}

\item{silent}{whether to suppress all console output.}
}
\description{
Run MCMC either with or without parallel tempering turned on.
  Minimum inputs include a data object, a data.frame of parameters, a
  log-likelihood function and a log-prior function. Produces an object of
  class \code{drjacoby_output}, which contains all MCMC output along with
  some diagnostics and a record of inputs.
}
\details{
Note that both \code{data} and \code{misc} are passed into
  log-likelihood/log-prior functions *by reference*. This means if you modify
  these objects inside the functions then any changes will persist.
}
